Return frozen parameters as fp32 from get_fp32_state_dict_from_zero_checkpoint
The optimizer keeps no fp32 master copy of a frozen parameter, so the checkpoint
stores it in whatever dtype the model held it in. Both merge paths hand that
tensor straight through, so a bf16 or fp16 run came back with a state_dict whose
trainable parameters and buffers were fp32 while every frozen parameter was not,
which is the common shape for LoRA-style training where the base model is frozen.

Cast in the eager branch of get_fp32_state_dict_from_zero_checkpoint, which is
the boundary that documents the fp32 contract. Lazy mode is left dtype-preserving
on purpose: convert_zero_checkpoint_to_state_dict consumes it and applies its own
--dtype, so upcasting there would both undo that option and cost a full fp32 copy
of the frozen weights during conversion. Casting an already-fp32 tensor with
Tensor.to returns the same object, so nothing is copied for the common case.

+++ deepspeed/utils/zero_to_fp32.py
@@ -646,7 +646,10 @@ def get_fp32_state_dict_from_zero_checkpoint(checkpoint_dir,
     if lazy_mode:
         return state_dict
     else:
-        return to_torch_tensor(state_dict)
+        # Frozen parameters come out of the checkpoint in the dtype the model held them in, since
+        # the optimizer keeps no fp32 master copy of them. Cast here so the whole state_dict is
+        # fp32 as documented, the same way buffers are already upcast in parse_model_states.
+        return to_torch_tensor(state_dict, dtype=torch.float32)
 
 
 def convert_zero_checkpoint_to_state_dict(checkpoint_dir,
